Refactor: extract feedback buttons to prevent duplication
- Extract user feedback thumbs up/down into an independent FeedbackButtons component.
- Move rendering logic to ChatArea to display feedback buttons only once per langfuseTraceId.
- Filter out function responses in chatCache to maintain correct message sequence for feedback rendering.
- Remove redundant copy buttons and Langfuse dependencies from AgentMessage.
