+++ src/components/ChatArea.tsx
@@ -1,6 +1,8 @@
 import { useEffect, useRef } from 'react'
+import React from 'react'
 import { UserMessage } from './UserMessage'
 import { AgentMessage } from './AgentMessage'
+import { FeedbackButtons } from './FeedbackButtons'
 import { ChatInput } from './ChatInput'
 import type { ChatMessage } from '@/lib/adk'
 
@@ -31,29 +33,25 @@ export function ChatArea({
         ref={scrollRef}
         className="flex-1 overflow-y-auto p-4 md:p-6 space-y-6 md:space-y-8 chat-container pb-0"
       >
-        {/* Date separator */}
-        {messages.length > 0 && (
-          <div className="flex justify-center">
-            <span className="text-xs text-text-muted bg-gray-100 px-3 py-1 rounded-full">
-              {new Date().toLocaleDateString('zh-TW', {
-                month: 'long',
-                day: 'numeric',
-              })}{' '}
-              {new Date().toLocaleTimeString('zh-TW', {
-                hour: '2-digit',
-                minute: '2-digit',
-              })}
-            </span>
-          </div>
-        )}
+        {messages.map((msg, index) => {
+          const nextMsg: ChatMessage | undefined = messages[index + 1];
 
-        {messages.map((msg) => {
-          if (msg.role === 'user') {
-            return msg.author === 'user'
-              ? <UserMessage key={msg.id} message={msg} />
-              : null // function response, or anything else; render nothing
-          }
-          return <AgentMessage key={msg.id} message={msg} />
+          return <React.Fragment key={msg.id}>
+            {msg.author === 'user'
+              ? <UserMessage message={msg} />
+              : <AgentMessage message={msg} />}
+            {
+              /* Show thumbs up/down when all below are true:
+                - Message has trace id
+                - Message is not streaming
+                - Next message has different trace id or doesn't exist
+              */
+              msg.langfuseTraceId &&
+              !isStreaming &&
+              (!nextMsg?.langfuseTraceId || msg.langfuseTraceId !== nextMsg.langfuseTraceId) &&
+              <FeedbackButtons traceId={msg.langfuseTraceId} />
+            }
+          </React.Fragment>
         })}
 
         {
